Part 6 Done!
https://www.youtube.com/watch?v=x7gxKmhcOfw

+++ Life Gauge TDD/Assets/Tests/Infrastructrue/ImageBuilder.cs
@@ -5,9 +5,9 @@ using UnityEngine.UI;
 
 public class ImageBuilder : TestDataBuilder<Image>
 {
-    private int _fillAmount;
+    private float _fillAmount;
 
-    public ImageBuilder(int fillAmount)
+    public ImageBuilder(float fillAmount)
     {
         _fillAmount = fillAmount;
     }
@@ -16,7 +16,7 @@ public class ImageBuilder : TestDataBuilder<Image>
     {
     }
 
-    public ImageBuilder WithFillAmount(int fillAmount)
+    public ImageBuilder WithFillAmount(float fillAmount)
     {
         _fillAmount = fillAmount;
         return this;
